Allages se vasi klassika kai merikes prosthikes sti leitourgia tou products

+++ backend/hit-it/src/main/java/com/hitit/models/Bidder.java
@@ -25,6 +25,6 @@ public class Bidder {
 
     private String location;
     private String country;
-    private Float rating;
+    private Long rating;
 
 }
